intend between teams added toggle view team body added project structure added
