+++ src/App.js
@@ -1,72 +1,11 @@
 import './Styles/App.scss';
-import {Component} from 'react';
-import teamIcon from "./images/team.svg";
-import toggleIcon from "./images/toggle.svg";
+import {Component, React} from 'react';
 import plusIcon from "./images/plus.svg";
 import PropTypes from "prop-types";
+import {Navigation} from "./components/Navigation/Navigation";
+import {Teams} from "./components/Teams/Teams";
+import {Modal} from "./components/Modal/Modal";
 
-// utilits_____________________________________________________________________________________________________________
-function checkVacationsDate(vacations, date) {
-  let result = false;
-  vacations.forEach((item) => {
-    const startDateNumbers = item.startDate.split(".");
-    const startDate = `${startDateNumbers[2]}/${startDateNumbers[1]}/${startDateNumbers[0]}`;
-    const endDateNumbers = item.endDate.split(".");
-    const endDate = `${endDateNumbers[2]}/${endDateNumbers[1]}/${endDateNumbers[0]}`;
-    if (date >= new Date(startDate) && date <= new Date(endDate)) {
-      result = true;
-    }
-  });
-  return result;
-}
-
-function Navigation({date, next, prev}) {
-  return <div className="calendarBar">
-    <button onClick={prev} className="calendarBar__nav calendarBar__nav_prev"/>
-    <span className="calendarBar__current">{date.toLocaleDateString("en-US", {month: "long", year: "numeric"})}</span>
-    <button onClick={next} className="calendarBar__nav calendarBar__nav_next"/>
-  </div>
-}
-
-function Teams({teams, allDays, date}) {
-  return teams.map((team, i) =>
-    <>
-      <tr className="calendarTable__team-header">
-        <td key={i}>
-          <div className="calendarTable__team-title">
-            <span className="calendarTable__team-name">{team.name}</span>
-            <span className="calendarTable__team-count">
-                  <img src={teamIcon}/>
-                  <span>{team.members.length}</span>
-                </span>
-            <span className="calendarTable__percentage">{team.percentageOfAbsent[date.getMonth()]}%</span>
-            <button className="calendarTable__team-toogle">
-              <img src={toggleIcon}/>
-            </button>
-          </div>
-        </td>
-        {allDays.map((day, i) => {
-            return <td key={i} className={day.isDayOff ? "calendarTable__dayOff" : ""}>
-            </td>
-          }
-        )}
-      </tr>
-      {team.members.map(((member, i) => {
-        return <tr key={i}>
-          <td>
-            <div className="calendarTable__team-title">{member.name}</div>
-          </td>
-          {allDays.map((day, i) => {
-              return <td key={i}
-                         className={`${day.isDayOff ? "calendarTable__dayOff" : ""} ${checkVacationsDate(member.vacations, day.fullDate) ? "calendarTable__vacations" : ""}`}>
-              </td>
-            }
-          )}
-        </tr>
-      }))}
-    </>
-  )
-}
 
 function Table({allDays, teams, date, modalToggle}) {
   return <table className="calendarTable">
@@ -97,104 +36,6 @@ Table.propTypes = {
   allDays: PropTypes.arrayOf(PropTypes.object),
   teams: PropTypes.array,
   date: PropTypes.object
-}
-
-class Modal extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      teamSelectValue: "Team name",
-      userSelectValue: "User name",
-      typeDayOff: "Type Of Day Off",
-      startDayVocation: "2021-02-22",
-      endDayVocation: new Date().toLocaleDateString("en-US", {year: "numeric", month: "numeric", day: "numeric"})
-    }
-    this.handleTeamSelect = this.handleTeamSelect.bind(this);
-    this.handleUserSelect = this.handleUserSelect.bind(this);
-    this.handleDayOffSelect = this.handleDayOffSelect.bind(this);
-    this.handleStartDayVocation = this.handleStartDayVocation.bind(this);
-    this.handleEndDayVocation = this.handleEndDayVocation.bind(this);
-  }
-
-  handleTeamSelect(e) {
-    this.setState({teamSelectValue: e.target.value});
-  }
-
-  handleUserSelect(e) {
-    this.setState({userSelectValue: e.target.value});
-  }
-
-  handleDayOffSelect(e) {
-    this.setState({typeDayOff: e.target.value});
-  }
-
-  handleStartDayVocation(e) {
-    this.setState({startDayVocation: e.target.value});
-  }
-
-  handleEndDayVocation(e) {
-    this.setState({endDayVocation: e.target.value});
-  }
-
-  render() {
-    // console.log("Start render",this.state.startDayVocation)
-    // console.log("End render",this.state.endDayVocation)
-    return <div className="modal">
-      <div className="modal__container">
-        <div className="modal__header">
-          <h2 className="modal__title">Vacation Request</h2>
-          <span className="modal__vacation-period">8 Days</span>
-        </div>
-        <form className="form" action="">
-          <h3 className="form__title">Dates</h3>
-          <div className="form__date flex-row">
-            <label className="form__date-label flex-column" htmlFor="start">
-              From
-              <input className="form__date-input" value={this.state.startDayVocation}
-                     onChange={this.handleStartDayVocation} id="start" type="date"/>
-            </label>
-            <label className="form__date-label flex-column" htmlFor="end">
-              To
-              <input className="form__date-input" value={this.state.endDayVocation} onChange={this.handleEndDayVocation}
-                     id="end" type="date"/></label>
-          </div>
-
-          <label className="form__title flex-column">
-            Team
-            <select className="form__select" value={this.state.teamSelectValue} onChange={this.handleTeamSelect}>
-              <option>Team name</option>
-              {this.props.teams
-              && this.props.teams.map((team, i) => <option value={team.name} key={i}>{team.name}</option>)}
-            </select>
-          </label>
-          <label className="form__title flex-column">
-            User
-            <select className="form__select" value={this.state.userSelectValue} onChange={this.handleUserSelect}>
-              <option>User name</option>
-              {this.props.teams
-                ? (this.state.teamSelectValue && this.state.teamSelectValue !== "Team name"
-                  && this.props.teams.find(item => item.name === this.state.teamSelectValue).members.map((member, i) =>
-                    <option value={member.name} key={i}>{member.name}</option>)
-                )
-                : null}
-            </select>
-          </label>
-          <label className="form__title flex-column">
-            Vac Type
-            <select className="form__select" value={this.state.typeDayOff} onChange={this.handleDayOffSelect}>
-              <option>Type Of Day Off</option>
-              <option value="Paid">Paid Day Off (PD)</option>
-              <option value="UnPaid">UnPaid Day Off (UPD)</option>
-            </select>
-          </label>
-          <div className="form__footer">
-            <button className="button button_b" onClick={this.props.modalToggle}>Cancel</button>
-            <button className="button button_a">Send</button>
-          </div>
-        </form>
-      </div>
-    </div>
-  }
 }
 
 class App extends Component {
@@ -297,7 +138,6 @@ class App extends Component {
     })
       .then((response) => response.json())
       .then((json) => {
-        console.log(json)
         this.setState({teams: json.teams});
       }).catch(err => console.log(err));
   }
