+++ src/Styles/blocks/_modal.scss
@@ -9,7 +9,7 @@
   z-index: 2;
 
   &__container {
-    margin: 20% auto 0;
+    margin: 100px auto 0;
     max-width: 393px;
     background: #FFFFFF;
     border-radius: 5px;
