+++ src/components/Navigation/Navigation.js
@@ -1,0 +1,9 @@
+import {React} from "react";
+
+export function Navigation({date, next, prev}) {
+    return <div className="calendarBar">
+        <button onClick={prev} className="calendarBar__nav calendarBar__nav_prev"/>
+        <span className="calendarBar__current">{date.toLocaleDateString("en-US", {month: "long", year: "numeric"})}</span>
+        <button onClick={next} className="calendarBar__nav calendarBar__nav_next"/>
+    </div>
+}
